chore: Refactor ContentPanel, detach tabs
