+++ packages/frontend/src/features/sqlRunner/components/ContentPanel.tsx
@@ -6,11 +6,8 @@ import {
 } from '@lightdash/common';
 import {
     Box,
-    Group,
-    Indicator,
     LoadingOverlay,
     Paper,
-    SegmentedControl,
     Stack,
     Text,
     Tooltip,
@@ -18,11 +15,7 @@ import {
 } from '@mantine/core';
 import { useElementSize, useHotkeys } from '@mantine/hooks';
 import { notifications } from '@mantine/notifications';
-import {
-    IconChartHistogram,
-    IconCodeCircle,
-    IconGripHorizontal,
-} from '@tabler/icons-react';
+import { IconGripHorizontal } from '@tabler/icons-react';
 import type { EChartsInstance } from 'echarts-for-react';
 import {
     useCallback,
@@ -42,14 +35,12 @@ import { ConditionalVisibility } from '../../../components/common/ConditionalVis
 import MantineIcon from '../../../components/common/MantineIcon';
 import { updateChartSortBy } from '../../../components/DataViz/store/actions/commonChartActions';
 import {
-    cartesianChartSelectors,
     selectCompleteConfigByKind,
     selectPivotChartDataByKind,
 } from '../../../components/DataViz/store/selectors';
 import { ChartDataTable } from '../../../components/DataViz/visualizations/ChartDataTable';
 import ChartView from '../../../components/DataViz/visualizations/ChartView';
 import { Table } from '../../../components/DataViz/visualizations/Table';
-import RunSqlQueryButton from '../../../components/SqlRunner/RunSqlQueryButton';
 import { useOrganization } from '../../../hooks/organization/useOrganization';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { DEFAULT_SQL_LIMIT } from '../constants';
@@ -59,32 +50,25 @@ import {
     selectActiveChartType,
     selectActiveEditorTab,
     selectFetchResultsOnLoad,
-    selectLimit,
     selectProjectUuid,
     selectResultsTableConfig,
-    selectSavedSqlChart,
     selectSql,
     selectSqlQueryResults,
     selectSqlRunnerResultsRunner,
     setActiveEditorTab,
     setEditorHighlightError,
-    setSqlLimit,
 } from '../store/sqlRunnerSlice';
 import { runSqlQuery } from '../store/thunks';
-import { ChartDownload } from './Download/ChartDownload';
-import { ResultsDownloadFromUrl } from './Download/ResultsDownloadFromUrl';
+import { ContentPanelTabs } from './ContentPanelTabs';
 import { SqlEditor } from './SqlEditor';
-import { SqlQueryHistory } from './SqlQueryHistory';
 
 export const ContentPanel: FC = () => {
     // State we need from redux
-    const savedSqlChart = useAppSelector(selectSavedSqlChart);
     const fetchResultsOnLoad = useAppSelector(selectFetchResultsOnLoad);
     const projectUuid = useAppSelector(selectProjectUuid);
     const sql = useAppSelector(selectSql);
     const selectedChartType = useAppSelector(selectActiveChartType);
     const activeEditorTab = useAppSelector(selectActiveEditorTab);
-    const limit = useAppSelector(selectLimit);
     const resultsTableConfig = useAppSelector(selectResultsTableConfig);
     const isLoadingSqlQuery = useAppSelector(
         (state) => state.sqlRunner.queryIsLoading,
@@ -227,11 +211,6 @@ export const ContentPanel: FC = () => {
         );
     }, [queryResults, activeEditorTab, currentVizConfig]);
 
-    const canSetSqlLimit = useMemo(
-        () => activeEditorTab === EditorTabs.VISUALIZATION,
-        [activeEditorTab],
-    );
-
     const resultsRunner = useAppSelector((state) =>
         selectSqlRunnerResultsRunner(state),
     );
@@ -239,8 +218,6 @@ export const ContentPanel: FC = () => {
     const pivotedChartInfo = useAppSelector((state) =>
         selectPivotChartDataByKind(state, selectedChartType),
     );
-
-    const resultsFileUrl = useMemo(() => queryResults?.fileUrl, [queryResults]);
 
     useEffect(() => {
         if (queryResults && panelSizes[1] === 0) {
@@ -251,14 +228,6 @@ export const ContentPanel: FC = () => {
 
     const [activeEchartsInstance, setActiveEchartsInstance] =
         useState<EChartsInstance>();
-
-    const hasUnrunChanges = useAppSelector(
-        (state) => state.sqlRunner.hasUnrunChanges,
-    );
-    const hasErrors = useAppSelector(
-        (state) =>
-            !!cartesianChartSelectors.getErrors(state, selectedChartType),
-    );
 
     const handleTableHeaderClick = useCallback(
         (fieldName: string) => {
@@ -304,152 +273,9 @@ export const ContentPanel: FC = () => {
     return (
         <Stack spacing="none" style={{ flex: 1, overflow: 'hidden' }}>
             <Tooltip.Group>
-                <Paper
-                    shadow="none"
-                    radius={0}
-                    px="md"
-                    py={6}
-                    bg="gray.1"
-                    sx={(theme) => ({
-                        borderWidth: '0 0 1px 1px',
-                        borderStyle: 'solid',
-                        borderColor: theme.colors.gray[3],
-                    })}
-                >
-                    <Group position="apart">
-                        <Group position="apart">
-                            <Indicator
-                                color="red.6"
-                                offset={10}
-                                disabled={!hasErrors || mode === 'virtualView'}
-                            >
-                                <SegmentedControl
-                                    display={
-                                        mode === 'virtualView'
-                                            ? 'none'
-                                            : undefined
-                                    }
-                                    styles={(theme) => ({
-                                        root: {
-                                            backgroundColor:
-                                                theme.colors.gray[2],
-                                        },
-                                    })}
-                                    size="sm"
-                                    radius="md"
-                                    data={[
-                                        {
-                                            value: EditorTabs.SQL,
-                                            label: (
-                                                <Tooltip
-                                                    disabled={!hasUnrunChanges}
-                                                    variant="xs"
-                                                    withinPortal
-                                                    label="You haven't run this query yet."
-                                                >
-                                                    <Group spacing={4} noWrap>
-                                                        <MantineIcon
-                                                            color="gray.6"
-                                                            icon={
-                                                                IconCodeCircle
-                                                            }
-                                                        />
-                                                        <Text>SQL</Text>
-                                                    </Group>
-                                                </Tooltip>
-                                            ),
-                                        },
-
-                                        {
-                                            value: EditorTabs.VISUALIZATION,
-                                            label: (
-                                                <Tooltip
-                                                    disabled={
-                                                        !!queryResults?.results
-                                                    }
-                                                    variant="xs"
-                                                    withinPortal
-                                                    label="Run a query to see the chart"
-                                                >
-                                                    <Group spacing={4} noWrap>
-                                                        <MantineIcon
-                                                            color="gray.6"
-                                                            icon={
-                                                                IconChartHistogram
-                                                            }
-                                                        />
-                                                        <Text>Chart</Text>
-                                                    </Group>
-                                                </Tooltip>
-                                            ),
-                                        },
-                                    ]}
-                                    value={activeEditorTab}
-                                    onChange={(value: EditorTabs) => {
-                                        if (isLoadingSqlQuery) {
-                                            return;
-                                        }
-
-                                        if (
-                                            value ===
-                                                EditorTabs.VISUALIZATION &&
-                                            !queryResults?.results
-                                        ) {
-                                            return;
-                                        }
-
-                                        dispatch(setActiveEditorTab(value));
-                                    }}
-                                />
-                            </Indicator>
-                        </Group>
-                        <Group spacing="xs">
-                            {activeEditorTab === EditorTabs.SQL && (
-                                <SqlQueryHistory />
-                            )}
-                            <RunSqlQueryButton
-                                isLoading={isLoadingSqlQuery}
-                                disabled={!sql}
-                                onSubmit={() => handleRunQuery(sql)}
-                                {...(canSetSqlLimit
-                                    ? {
-                                          onLimitChange: (l) =>
-                                              dispatch(setSqlLimit(l)),
-                                          limit,
-                                      }
-                                    : {})}
-                            />
-                            {activeEditorTab === EditorTabs.VISUALIZATION &&
-                            !isVizTableConfig(currentVizConfig) &&
-                            selectedChartType ? (
-                                <ChartDownload
-                                    fileUrl={
-                                        pivotedChartInfo?.data?.chartFileUrl
-                                    }
-                                    columnNames={
-                                        pivotedChartInfo?.data?.columns?.map(
-                                            (c) => c.reference,
-                                        ) ?? []
-                                    }
-                                    chartName={savedSqlChart?.name}
-                                    echartsInstance={activeEchartsInstance}
-                                />
-                            ) : (
-                                mode === 'default' && (
-                                    <ResultsDownloadFromUrl
-                                        fileUrl={resultsFileUrl}
-                                        columnNames={
-                                            queryResults?.columns.map(
-                                                (c) => c.reference,
-                                            ) ?? []
-                                        }
-                                        chartName={savedSqlChart?.name}
-                                    />
-                                )
-                            )}
-                        </Group>
-                    </Group>
-                </Paper>
+                <ContentPanelTabs
+                    activeEchartsInstance={activeEchartsInstance}
+                />
 
                 <PanelGroup
                     direction="vertical"
